Moving components to specific folder
