+++ src/components/Admin/EntryCard.vue
